Add cursive termination upon running the game
This commit introduces some unsafe code that sets the exit condition of the Cursive instance. It could probably be improved by using a Mutex with the tuple inside to prevent simultanious access, but currently the reads and writes happen in situations that could never coincide in the program's execution. They are as follows:

* When the program starts
* When hitting the play button (This is the only time the condition is set during the Cursive's existence)
* After the cursive has been dropped