Added hunt support
The ulimit in the command is due to an issue with terminating hunt. If someone
hits ^C, it kills the login screen application, but hunt will keep trying to
read from an stdin that does not exist anymore, consuming as much CPU as
possible.  The solution to this is a bit of a hack, since it just kills the
process if it takes more than 5 seconds of CPU time, which it shouldn't do in
normal operation.